fix handling of spaces
This is about:
close #42
close #34

This includes changes for onyo new (mentioned in the issue),
and updates also all other commands for the usage of spaces
in file/folder names, too. It also has updates and changes
in different test cases to include folders/files with spaces
("user2" -> "user 2"). To allow for escaped characters in
tests, the .flake8 got updated.