rm: move file checks to beginning of call, make command robust
robustness means:
`onyo rm <folder> <folder/file>`
Does not throw an error, because check in the beginning was
successful, and at deletion of second element it just continues,
because it is assumed the element disappeared in the same call.

`onyo rm <file> <file>`
Similar, this should not error, because when all files to delete
that were given and checked are fine, `onyo rm` can assume it is
still fine. This might be especially relevant to not error when:
`onyo rm <folder/file1.txt> <...> <folder/*>`