edit: vi as fallback; error handling and check if installed
Before this commit: either $EDITOR was set, or nano was used as
fallback without further checks.

Now: Adds vi as fallback, but also checks if editor was set to
empty string ($EDITOR= ), instead of unset. Also checks for wrongly
set editor ($EDITOR=asdf), and if nano/vi are available if used.