all: remove search for rel. path from outside onyo, when environmental is set
This commit removes the functionality to check first the environmental, then
the local directory as a possible onyo dir. Like git, when onyo is called
from outside of the directory, without having set explicitly the environmental
variable, it will not try to use the relative path from the arguments anymore,
but just throw an error. If no environmental variable is set, onyo will use
the current working directory as the onyo root directory. Instead of finding
the onyo dir in each function depending on the arguments, now the onyo  dir
is decided in the onyo main.py function.

This change might make the usage of onyo more stable, and the code got strongly
simplefied. Also, the implementation and call for ´onyo fsck´ in the future
will be simpler, when the directory which to verify is already decided on the
top level.

Additionally, this commit removes the test cases for "check commands with
relative paths outside of the onyo directory", since the functionality does
no longer exist.