new: check not just if asset name exists in folder, but anywhere in onyo
Before, `onyo new` just checked for the existence of the new text-file
it wanted to create. Now it uses the function in utils to check if an
asset already exists anywhere in the onyo repository (except if the
other file is in .gitignore. If an ignored folder/file exists like:
ignored_folder/I_am.txt
and we later allow different name schemes, it would be still possible
to create an asset with that name. This is open to discussion).