fsck: move code that verifies different aspects of onyo into separate functions
Depending on the way that onyo is called, it will be wished to check
different aspects of the onyo repository. For now I have the function
that verifies all aspects (the explicite call of `onyo fsck`) and another
function read_only_fsck(), that just checks for being in a onyo repository
and the validity of the yaml structure of the asset files.

This is a dummy differentiation to check the code; it enables to run the
full checks for all commands, and run a part of the checks for read-only
commands. It is with this change possible and very easy to build for
example a separate fsck-function for each of the onyo commands, if wished,
or to add further validity checks to fsck and decide, when they should be
run. Specifities can be discussed at any time.