all commands: check validity of repo with onyo fsck before running any command
Add before all commands some form of onyo fsck (diffentiates currently
between read-only or commiting commands, can later be further ajusted).

The only exception is ´onyo init´, where no onyo fsck will be run, but
instead just the normal checks already discussed (e.g. exists path, etc).

´onyo git´ makes always a full check, even if the git command might be
a read only command, because it is hard to differentiate between both.
We might want to change this later to a special case, where it runs the
tests, but does not error, but just warns in case of invalidity of the
repository, so that ´onyo git´ can be used to repair the repository.