edit/utils: check yaml syntax before saving of changes in asset
This creates/edits a temporary file. If the changes mess up the syntax,
the user is asked if he wants to further edit. If yes, temp file is
opened again and the wrong syntax can be corrected, otherwise the temp
is deleted.

If a temp file exists while calling onyo edit to begin with, the user
is given the option to edit it, otherwise it will be overwritten with
a new temp file.
Currently, for this to work, .onyo/temp has to be in .gitignore,
otherwise the fsck will mention the broken syntax before opening the
file and not allow the onyo command. Either the temp folder should be
ignored, or a separate type of fsck be done just for edit.