history: use onyo config variables (e.g. history.interactive)
This removes the flag, instead uses always the variables in onyo
config. For this, `onyo init` will set the values in onyo config
to default fields. Updates the error messages and other related
stuff in onyo history.

This could all have been part of the initial `onyo history` commit,
but since I added config afterwards, I decided to have it this way.