templates: add default template, add it to MANIFEST.in and setup.py
Add the first default template (like specified in readme with fields
and reserved characters) and add it to manifest and setup.py

This template is just defining what is currently already happening
in a hardcoded way, but it will also be a blueprint for other templates
later.

The changes in manifest and setup.py make sure that when installing
onyo with pip (like I do currently) the template(s) will be part of
the package in the end, too.
This should later make it possible to add "default" templates to onyo
(e.g. laptops and other devices) and copy them into every onyo repo when
calling onyo init, or possibly even "repair" the template folder of an
onyo repository if somebody breaks it or deletes a template or something
like that.