templates: add --template to `onyo new`
adds the flag in main.py, and adds to `onyo init` the creation
of a folder ./onyo/templates/ (and anchors it) were all templates
from the onyo project are copied into.
The templates are currently just "default".

This commit updates the command `onyo new`, which now uses the
the template flag (if not set, it uses the default template).
It reads then the name scheme, field names, and reserved characters
from the template, as well as value fields, which will then be
written into the newly created asset file.

This is written in a way, that nothing about the current behaviour
of `onyo new` is changed at all, it can be completely used like before:
> onyo new shelf
> onyo new --template default shelf
Do the same, it just enables easy adding and using of new templates.