readme: add explanation to templating, examples, commands
The sections about templates, reserved characters and name scheme
were mentioning parts of onyo that are currently solved involved in
templating.

The command `onyo new --template <template> <dir>` is expanded
with a description for the template flag.

Additionally, example templates are described in the end of the
readme. Even if it is just describing the current way it works
(which will very likely be modified in the future), I think it is
helpful to provide easy examples, so that future discussions and
tests can be about concrete use cases.