new/templating: copy templates instead of defining fields in them
The first form of templating in onyo happened with the templates
defining the name, fields, and other things.

Now the name is back to the default, the name fields are pre-defined,
and the templates will simply be copied and then opened with the
editor. The changes are based on discussions/meetings, and other
functionalities (more flexible names) will be added later.