validation: add dummy rules, update onyo init and MANIFEST.in
As proof of concept I added a dummy rule file to onyo/validation
and to the manifest.in file. The validation folder and its files
get copied with the installation of onyo.
I also updated `onyo init` so that it creates a folder
onyo/validation/ and copies the validation files into it.

This is meant to provide a base on site of onyo where we can
later add more sophisticated rules and validation files.
It helps while programming to have some mininum rules-file,
and onyo site of onyo init it makes sure that later usage will
be comfortable.