validation: use validation file to check syntax after new/edit
The validation file added with the last commit is now used when
creating new files with `onyo new [--template] <dir>` and onyo
edit and warns the user if some changes made do not fit the rules
specified in the validation file.

Currently the mechanism just checks for the type and just allows
int and str.