new: fix bug and update message when new asset not validates
There was a bug, that when a newly created file does not validate,
it would still create an asset and not add it to git. This was
fixxed in new/utils.

Additionally, when a new file not validates, and the user does
not want to correct the error (with "further editing?"), a specific
message will be printed, that tells the user not just that no changes
were made, but that actually no new asset was created.