new: add faux serial number generation
When for the serial field "faux" is added, it generates a faux serial
number, that looks e.g. like:
laptop_apple_macbookpro.faux78d7eb78, the current length of the faux
number is 8, but can be changed with a parameter in case later something
different or more flexible is wanted.

Close #111