validation and templates: rename or comment out examples
The example template got renamed `laptop.example`, so the user
can see how a template for an asset might look like, but still
create a new one called laptop without over-writing the example.

The validation.yaml contained a lot of examples from the readme,
these got commented out, so that a user can create a new onyo
repository, and create new assets in there without having to
worry about failing the validation because of the choosen examples.

Close #104