set: update output and file list in command with rel. or abs. path
At two places, the commands were assuming that onyo would run in the
repository, and did not handle relative vs. absolute paths right.
This updates the list of files to copy (which fixxes a bug, since
running onyo set -C created some problems) and updates the diff-like
output, so that file names are displayed always relative from asset
to the onyo root.