tests: add test for onyo cat; set: order list of files to update
Onyo cat was so far untested, since the asset files were always
empty. With the new onyo set tests, this command can be tested,
too. I also added another onyo set test case.

The change in `set` makes sure, that the files get always changed
in the same order (alphabetically). This is mainly necessary for
the tests to work, but I could see this be helpful in general
and possibly good for the other commands too.