shell-completion: new command
This adds the core of what is needed for tab-completion in bash and zsh
shells.

It's written using bash's completion functionality, and will work for
zsh through its compatibility support.

Numerous corner cases are not supported, most especially around
file/directory completion, `onyo git`, and limiting the number of
argument accepted.

The majority of the script parses the output of `onyo --help`
(or onyo <command> --help), to keep maintenance relatively low.

Fixes #28