edit: arg var file->asset
I chose not to rename "sources" to "assets". it would change most lines,
and there was already a file->sources thing going on. The current
situation is not ideal, but it's not any worse.

This code will likely be revisited at some point anyway, and the cleanup
can happen then.