utils: neuter type checks
So.... the path checking works for everything except when onyo -C is
used. And I see no way to pass the information required to make that
situation work. ArgParse sends only a single string (the argument) to
the function/class. [1]

So these "checks" are now no-ops. And just make life possible for the
tab-completion scripts.

I'm intentionally choosing to not squash this commit, on the off-chance
that someone has a better idea, and can reuse some of this approach.

[1] https://docs.python.org/3/library/argparse.html#type