mkdir: refactor
This reduces the number of functions from 5 to 3, adds docstrings, and
overall simplifies the logic and var names.

Ideally the dir names would be all relative to the root of the onyo
repository. A TODO has been added about this.

It would also be more performant to build a list of anchors and add them
all in one call, rather than invoke a new process for each add. But I
don't see people adding 10k directories at once, so I figured it best to
leave the code simple as it is.