init: refactor
This reduces the number of functions, adds docstrings, and greatly
simplifies the logic.

The biggest change is the use of a `skel/` directory (done in the
previous commit). This simplifies the creation of the `.onyo` dir, as it
no longer needs to be created programmatically, and instead can simply
be copied in whole.

Fixes #138