TST: add code coverage
This was more complicated than I expected.

Coverage is complicated by the face that the tests use "subprocess".
Thankfully "coverage" has this well documented:
- https://coverage.readthedocs.io/en/latest/subprocess.html

The second issue is that coverage is executed in relation to the current
working directory. Onyo's tests change the CWD. A solution/workaround
(and a digression into the difference between the two) is detailed here:
- https://github.com/nedbat/coveragepy/issues/1093

Lastly, because the tests are invoking the command "onyo', it was not
executing the code in the source repository, but rather at the install
location. This is solved by doing `pip install` in editable mode:
- https://packaging.python.org/en/latest/guides/distributing-packages-using-setuptools/#working-in-development-mode