TST: add tests for `onyo new`
Since bigger changes in `onyo new` are planned, this commit
is mainly testing that the creation of assets generally works
in preparation of testing and changing functions from utils.py
like edit_file().

Runs `onyo new --non-interactive` with and without faux serial
number for a set of different folder names (e.g. with and without
spaces in folder names, different levels of depth).