RF `utils.get_list_of_assets(): use git index for listing assets
Lists first all files from the git index (to remove the internal
calling of git commands) and selects the ones which are not protected
by onyo (e.g.: .anchor, .git, .onyo) for returning.

Adds also a doc string to the function.

TODO: I still return [x[0][0], Path(x[0][0]).name] instead of just
x[0][0] for the asset list, because I have the unpatched version of
`onyo fsck`, this must be updated after merging the one of the PRs.