RF: `onyo edit`
`utils.edit_file()` and `utils.get_editor()` got moved to
`onyo edit`, and the code of `utils.edit_file()` got simplified.
The tests for the editor got moved to `test_edit.py`.

The commands `edit` and `new` are updated to adapt to the changes,
since the commands and functions now expect e.g. pathlib objects, use
the repo class, and the handling of problems changed.

`test_edit.py`:
- move tests for editor from `utils.edit_file()`
- add `test_edit_single_asset()`
- add `test_edit_multiple_assets()`
- add `test_edit_protected()`
- add `test_edit_without_saving_changes()`
- add `test_edit_non_existing_file()`
- add `test_continue_edit_no()`
- add `test_edit_without_changes()`
- add `test_edit_with_dot_dot()`