NF: add `git status`-like properties to Repo
This adds three properties to Repo:
- files_changed
- files_staged
- files_untracked

I was unsure whether to expose these as methods or properties. The
former makes it clear that there is no caching, and a cost. The latter
is more consistent.

In the end, I chose properties, as I feel it matches other information,
and also acts a stable API that we can update (or cache) later as
needed.