add: onyo unset and tests for onyo unset
The command and its tests are written to mirror the workings and
tests of `onyo set` as much as possible, and to just add on top
of it where differences require it (e.g. not allowing to remove
name-fields with unset, and test that an error is thrown, instead
of testing the onyo set functionality for updating them).