tests: remove `test_invoking.py` and `reference_output/`
These tests do not really add something to the process of writing onyo
at the moment, but need a lot of separate updating in between, because
their functions differ. To simplify coding, these are for now removed
and tests with a similar functionality but more modern stile should be
added later again.