set/unset: remove flags --recursive
The flag was in the beginning meant to stop the user from accidently
select a folder and recursively change multiple files. Since we now
have --yes flags for these commands, the user is asked to verify
changes anyways, and the flag became unnecessary. The limiting of
recursion can still be done with the --depth option, which works like
before.

This removes the --recursive flag from both commands, from the parsing
of arguments, the onyo repository class, and from the tests. Since
before there were separate tests to test the behaviour with and wihtout
the flag, there are now fewer tests then before.