BUG: make order of files/dirs in commit messages predictable
This makes the commit messages reproducible, which is necessary for #263

I considered changing Repo.files_staged to a list. However, other
properties would also need to be changed, otherwise it would be
inconsistent.

Furthermore, I felt that a list would be misleading, as the order
returned by git is merely alphabetized, rather the order in which files
were staged.

Thus, I decided to keep the properties as `set`s, and make the change
where it's specifically needed: the commit message.

fix #274