tree: use `subprocess.run()` instead of `os.system()`
For many reasons, we use `subprocess.run()` if possible, which is
the case for `onyo tree`.

`os.system()` leads to errors when used with certain characters,
so while overhauling the parsing of paths, we found that subprocess
works better for the `onyo tree` command.

Also updates the return-code and error-handling of tree, and removes
the changing of directories, since this all is unneeded when
`subprocess.run()` is used.