RF: modernize `test_mkdir.py`
This is mainly a modernization of the existing mkdir tests, which
uses and adds fixtures, doc-strings and other modernizations.

These changes shorten the existing code, and make it way easier to
test the same list of directories in many situations, so that this
refactoring also covers some additional situations, e.g. tests with
special characters, or more folders created in one call.