TST: update handling of directories to test for in `test_new.py`
`@pytest.mark.repo_dirs(*directories)` is now used for more flexible
setting of directories, and `./` was removed from the list of directories.
Creation of assets on the toplevel is instead a separate test case.

This commit also removes the double casting (which became unnecessary
because of the change above):
`assert str(Path(asset)) in ret.stdout` -> `assert asset in ret.stdout`