unset/argument parsing: use new syntax for command unset
Implements the new syntax from issue #286 for `onyo unset` (in parallel
to the changes for `onyo set` from one of the previous commits).

Also updates `lib/onyo.py`, because the parsing of a comma separated list
is there not necessary anymore (and I think should perhaps not have
happened there in the first place).