demo: update `reference_git_log.txt`
The changes in onyo main for the set command lead to integers being
used as integers e.g. `onyo set --keys USB=2 path [asset]` does add
`USB: 2` and not `USB: '2'` to the yaml asset.
This changes the hashes in the reference log, which is updated with
this commit.

The problems with datatypes and yaml are known, but should be tested
for specifically, especially as soon as `onyo new`s syntax is updated
too and `onyo get` is merged.