TST: update tests for `onyo new` to use new CLI syntax
This updates the tests to call `onyo new` with the new flags `--keys`
and `--path` to unify its syntax to follow the syntax of `onyo set`
and its flags.

Adds a test that uses `--keys` to overwrite contents of a template.

It changes also output and variable names from "set" to "keys" in
`new.py` and tests.