NF/mkdir: add flags --yes, --quiet; update "user response"; add tests
The main purpose of this commit is to normalize the behavior of
`onyo mkdir`. It is the only committing command that did not have the
flags `--yes` and `--quiet`, and this commit adds these to flags and
tests for them. Updates the `onyo mkdir` lines in the demo, too.

The user-response is taken from onyo edit (in the same way `onyo new`
does this) to normalize the user-response-request behavior, too.
While this will create a need to update it later, when we
move the interactive functions to another place, I still think this
commit will prepare the move, since just the import line has to be
changed later then, not the code of the mkdir() function itself.

The existing tests are updated (including lib tests that sometimes
use mkdir), since they have to use the `--yes` flag now, and the
`assert ret.stdout` lines check now that changes are mentioned in
the diff output.
At this point I removed some unneeded tests (e.g. 'overlap/three'
adds nothing new, if the folders 'overlap/one' and 'overlap/two'
already exist). I think we should remove unneeded tests like this
to make room for tests with special characters, and to have more
overlap between the functionalities of the commands and the API.