TST/RF: modernize `test_rm.py`
The main changes are, that I streamlined the tests a bit, so that they
begin with single file tests, then list of files, etc. The biggest
functional difference is, that I added tests that try to delete single
and many folders. Also, I updated a few of the doc strings, so that
they display fuller information, like in many of the other tests.