TST/RF: modernize `test_config.py`
I honestely think this will never change anything. Never the less,
all tests got the repo-object passed into, so I added a final fsck,
which checks that the repository stays in a clean state, like we do
for most of the other tests.
Also, I added pyre, because it bothers me, that we do not use it
everywhere possible.

These tests feel like they already cover a reasonable amount of
cases, when keeping in mind that config is meant to be very flexible.
Further updating will probably occure, when we think more about
what else to have configurable. For that moment in time, this file
is now a little bit modernized, too :)