TST/RF: modernize `test_init.py`
The usual modernizations. This test file did not get much attention
in the past. I added `assert`s to verify the output in correct and
incorrect cases. This is mainly meant to simplify futher changes in
the init tests.

onyo init prints at success into stderr (#305). As soon as the
logging overhaul of #255 is implemented, this test file has to be
updated to reflect the change.