NF/BF: clear caches after modifying repository
The function to re-set the caches to `None` do exist since the
implementation of `onyo set`, yet they were not used in most
functions, and neither in a consistent way.

This commit adds a `clear_caches()` call to the end of `Repo.add()`
and all functions in the API that modify the caches with a
git-subcall that does not use `Repo.add()` (e.g. `Repo.mv()` uses
a "git mv" internally and does not call `Repo.add()`, so it must
be used manually).

Additionally, this function now allows flags to deactivate the
resetting of caches.
The commands mv, new, rm, and especially `Repo.add()` can (or
should in the future) all be able to change/add/remove directories,
assets and templates, and thereby have to clear them always.

While doing these changes, I noticed how much we should have
something like API-workflow tests, or like the demo, since it is
quite hard to verify that not some other functions (e.g. `commit()`
as git function) also would need a clearing of caches.
I think these should not require this currently, because they get
mainly called by the command-functions, or in the command-files, but
we should add these tests soon anyways to be sure.

Close #248