BF/TST: check PATH is in repo for `onyo tree PATH`; add `test_tree.py`
`onyo tree PATH` did test that the opdir is in the repository, but
did not test if the PATH given as input is in the repository, which
I fixed.

To verify that the fix does not introduce new bugs, I added
`test_tree.py`, because it seems we never had any tests at all for
this command.

Close #121