RF: add some type hints
This contains the type hints added by me and `pyre infer -i`,
some new docstrings, and moves definition of variants into
parameterization (to stabelize random test order).

The goal of this is to go closer to `pyre --strict` requirements
and conform with our own "Code Conventions".

changes. Lines starting # with '#' will be ignored, and an empty message aborts the commit.