typing: add typehints generated by `pyre infer -i` to `tests/`
This adds the automatically generated type hints to the tests.
It mainly adds typehints to declarations of global variables.
In principal, we want to move this into the parametarization
itself, instead of declearing variables outside of the function,
but as a first approach and to share a part of what `pyre --strict`
would require us to do, I added it here.