NF: add `Repo._find_root()` to find root from subdirs inside a repo
Change function `Repo._get_root()` that requires a Repo-object to
`Repo._find_root()` that identifies the root of an existing
onyo repository from any subdir inside a repository.

This is the first step to change `Repo`s so that they can only be
instantiated with the root of an existing onyo repository and retiring
`Repo.opdir`.