get: update sanitizing and printing of paths to work with absolute paths
The sanitizing/validating of the `--path` arguments was updated to use
`repo.sanitize_path()` to use absolute paths in `get.py`.

By nature, get prints a bunch of paths, these were adjusted to be
relative to the `opdir`.

be
ignored, and an empty message aborts the commit. # # On branch pathing # Changes to be
committed: # modified:  onyo/commands/get.py # # Changes not staged for commit: # modified:
tests/commands/test_get.py #