TST: add tests for `Repo.generate_commit_message()`
Verify that all files staged for commit are mentioned in a
generated commit message, and that the header contains correct
information (modifiable "command name", number of changed files)
and the body all files changed.

Especially since switching to absolute paths, it is necessary to
verify that just paths relative to repo-root are in the commit
message.

Close #297