RF: Extract non-git implementations from Repo class
This is the first step in an effort to refactore the `lib` module.
It extracts the implementations of all methods that are not obviously
about plain git operations from the `Repo` class into either `utils.py`
or `command_utils.py` depending on whether comments and method naming
suggest it's written for use with a particular command.

This replaces all such methods' bodies by a local import of a function
with the same name from the respective utility module, thereby avoiding
circular imports at module level. Hence, possible merge conflicts should
be easy to resolve, since there's no change to the actual implementation
yet.

There are several iterations of reshuffling the extracted code being
worked on. The aim is to keep conflict resolution minimal by separating
things out in this set-by-step fashion.