RF: Remove artificial dependency on Repo
After the functions in `command_utils.py` being extraced from the `Repo`
class, they still called each other via an instance of `Repo`. Change to
direct calls, in order to remove an apparent but not real dependency on
that class. The only usage remaining should be calling methods that are
directly about git operations and using `Repo`'s properties.