Integrate tests in package
This patch moves tests into the onyo package's source tree, starting to
use an interleaved layout, where unit tests for every source code file
are to be found at the location of said file under
`tests/test_whatever.py`. Highest level tests are located under
`onyo/tests`. For now, this includes the tests for `main` as well as the
test for demo generation (which is testing a script that itself is not
currently part of the package).
Not only do I think, that it's easier to handle to have tests right next
to the source code they are supposed to test (including relative imports
allowing to move code with its tests w/o getting into import hell), but
somewhat more importantly: I'd consider the apparent idea to have tests
be a separate package an anti-feature for a tool like onyo. That is
because a lot of the details of its behavior depends on filesystem,
installed git, configurations, etc. on a given system. Hence, a
dedicated command performing (a subset of) the tests on user's system is
something we'd want to have eventually. This would re-use tests for
slightly different purpose: Probing a given system/installation for
"does onyo behave as expected here?" enabling the user to discover (and
properly report on) issues before running into possibly deceptive
manifestations of them amidst a workflow.