Implement CLI usecase test
Addressing the first part of #345, this patch implements a CLI workflow
outlined in issue #345.
There are not a lot of assertions in this test yet. There's little use
in repeating assertions from the command tests. This is supposed to

- test that desired operations are actually expressible with onyo's
  interface (not the currently the case for querying assets for a value
  within a given range and for limiting the matches to exclude certain
  subtrees)
- test that an entire flow doesn't encounter barriers that are
  unaccounted for in isolated tests for single commands
- test that and how information retrieved by one command can be used in
  the next
- inform further refinement of interfaces, features and outcome
  definitions

Additional assertions should come from decisions made (documented in
issues), that more precisely define what is the desired behavior. This
is particularly true for how the history should actually look like in
order to facilitate an easy assessment of what exactly happened.