Remove superfluous CWD specification for git calls
`GitRepo._git` uses `self.root` by default now. Furthermore, it makes
little sense to call git from elsewhere in e method that is bound to a
concrete repository. Hence, the only currently valid use of the `cwd`
argument is the git-init call.

Remove all other usages of `cwd` as well as `-C ...` in the command
call.

Closes #348