Don't stage in Onyorepo.mk_inventory_dirs
This patch is working towards a resolution of #370. Remove staging from
within `OnyoRepo.mk_inventory_dirs` and return actually added (and to be
staged/committed) files instead. As described in #370, in order to
resolve it, an incremental disentanglement of staging and the widespread
do-modify-but-possibly-rollback-approach is required.
Hence, removal of the staging here does not yet come up with a change to
when to modify the worktree. Therefore, figuring out what was actually
modified and return those paths is currently necessary to not break
reporting and rollback.