docs: add some minimal example usage to all commands
This, again, is not the end of all wisdom. I just started to add
at least one example use case to all commands, so as to demonstrate
the usage of it, which seems to me especially relevant to show e.g.
which commands have flags like --path, --keys, and for commands like
`onyo init` to make easily clear, that when no path is given the cwd
is used, or if a non-existing path is provided it gets set up, and
such things.