main.py: Normalize description of command/flags/parameters
This does a lot of minor changes and normalizations for all texts
specified in the arg-parser stuff. I do NOT move the argparser stuff
into the commands (we discussed this, but I did not make it work, so
I decided for now to just update the texts in main.py).

I use `help=textwrap.dedent(commands.cat.__doc__)` for all commands.
Before we just used this for the description of commands, and `help`
was instead a short sentence describing an argument. But like so often,
these small sentences were worse, out of date, and generally not good.
As long as we do not make an actual decission how these should look
like (and finish the design decissions for all commands) I decided to
just use the same text in both locations, so as to be at least
consistent and have the correct information in just one location.

Otherwise, I normalized a bunch of stuff, like the quotes used, when
and were to use brackets, capitalized letters, and how to write the
user-visible texts in general.

Like so often, we will have to go over these texts again, this commit
is just the first step, and meant to unify texts along all commands.