main.py/commands: remove/normalize arguments without value
Works towards #341. For different arguments/flags we used everything
between `None`, `[]`, `[.]` and probably others to declare values as
being empty/None by default.

This is bad while programming, and even worse because we display those
default values in the docs, were they just look terrible and are also
confusing.

For know I decided to remove all these different declarations in main.py,
and move those values to the locations in the commands (were we want to
do the arg-parser definitions later anyways), because there we sanitize
the different args anyways (and sometimes even check for None-values
already).

We might decide later to define default values (especially empty ones)
differently again, but for know this makes it at least consistent along
the different commands.