TST/DOC: add test and docstring to `OnyoRepo.generate_commit_message()`
Gives the function a new docstring in the style discussed in issue psyinfra#339.
The test modifies an OnyoRepo with the python implementations of the
commands and then calls `generate_commit_message()` and checks the
default commit message that was generated.

There is the idea to change the commit message body in a future PR.
This would impact parts of the test function written now, and I assume
the current test should help as a starting point to test the new format.