New inventory abstraction
This patch is the basis for further refactoring/rewriting.
It introduces an `Inventory` class and the concept of inventory
operations.
An `Inventory` is meant to be the main python interface for onyo.lib.
Command implementations (as well as arbitrary python users, of course)
are meant to operate on an `Inventory` object.
Anything an onyo command can do, boils down to basic operations on an
inventory with to types of entities an inventory knows: Directories and
Assets. A directory is represented by a `Path`, an Asset is expected to
be a dictionary (or at least a dict-like object).
Basic operations are creating, removing, moving, renaming of assets or
directories and modifying an asset's data. Hence, there are 9 such
operations.
The `Inventory` class comes with methods to "register" operations, such
as `add_asset()` or `move_directory()`. These methods generate the
required `InventoryOperation` objects and register them in a list
(`Inventory.operations`). At this point nothing is executed yet - the
operations are pending. Execution happens only upon
`Inventory.commit()`.
The `commit()` method will not only execute and git-commit what needs to
be done, but also generate a footer in the commit message that records
the basic operations done in that commit. This is somewhat different
from the information already contained in a git commit, because of
semantics and the fact that a directory is an actually relevant entity
in the inventory (as opposed to git). Hence, from the commit itself it's
hard to tell, whether assets where moved or an entire directory. But
this is different from an inventory point of view. Recording it like
that should allow for audits that are unaware of git, a relatively
straight-forward implementation of a history command that focuses on an
inventory view rather than a git view that needs interpretation by a
user and for things like a migration of the entire history to a
different inventory system.

Since onyo commands want to tell the user what would be done by a this
command, asking for confirmation before actually changing anything,
`Inventory.diff()` gives a git-diff-like view of what the currently
pending operations would do. In opposition to previous onyo versions,
this is done entirely in memory rather than first executing things, then
querying git for a diff and finally trying to rollback changes, if the
user didn't confirm.

An `InventoryOperation` is broken down into an `InventoryOperator` and a
tuple of operands. The `InventoryOperator` references three Callables
that implement an executor, a differ, and a recorder.
An executor's job is to actually perform the operation. A differ returns
a diff for a given operation (used by `Inventory.diff()` to assemble a
diff for all of the pending operations) and a recorder generates the
necessary snippet for the operations record in a commit message.
Operands can obviously be directories and/or assets, depending on the
operation. The low-level executors, differs, and recorders do no input
validation whatsoever. This is intentional, since the `Inventory`
methods passing the operands are supposed to already do that to avoid
partial execution later on. Running something like `path.exists()` over
and over at every abstraction layer seems like an unnecessary burden on
performance.