Rewrite most commands to use Inventory object
This is the first huge chunk of rewriting onyo commands based on the new
`Inventory` class and the concept of inventory operations.
Essentially untouched commands are `unset`, `get` and `fsck`. Some
others are not entirely there but supposed to be functional.
Furthermore `fsck` is removed from all other commands as well as the
tests. Re-introduction is intended but depends on a rewrite of this
command.

The `onyo new` command comes with a substantially changed interface.

In addition, most commands now fail on first occasion rather than
collecting everything wrong with their inputs they could possibly report
on. This is done to get the command logic straight, not because it's an
aim by itself. We may well re-introduce this behavior (but implemented
differently and hopefully consistent across commands).

Next to no effort has been put into documentation yet, which is expected
to be either missing or out-dated at this point.

Lastly, a couple of tests are failing with a `RuntimeError("TODO ...`
hinting at things that need to be decided before fixing this.