Allow onyo-rm to recursively remove directories
`onyo rm` now does remove non-empty directories, recording implicit
operations for its inventory contents.
In addition, this path prepares `onyo rm` for dealing with non-inventory
files (`.onyoignore`) and asset directories.
For non-inventory files, this relies on introducing sort of a dummy
`InventoryOperation` that does not produce a record in the commit
message, but is in-line with other operations with respect to pending
operations and `Inventory.diff()`.